requirements.txt
docutils==0.13.1
lockfile==0.12.2
python-daemon==2.1.2
wheel==0.24.0
